# frontend/openapi.json
{
 "components": {
  "schemas": {
...
   "FloatType": {
    "enum": [
     "cash",
     "emoney"
    ],
    "title": "FloatType",
    "type": "string"
   },
...
   "GeneratedBy": {
    "enum": [
     "llm",
     "template",
     "replay"
    ],
    "title": "GeneratedBy",
    "type": "string"
   },
...
   "GuardResult": {
    "enum": [
     "pass",
     "numbers_fail",
     "injection",
     "schema_fail",
     "timeout",
     "error"
    ],
    "title": "GuardResult",
    "type": "string"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Lang": {
    "enum": [
     "bn",
     "en"
    ],
    "title": "Lang",
    "type": "string"
   },
...
   "LlmIntent": {
    "enum": [
     "copilot",
     "narrate",
     "agent_briefing",
     "distributor_briefing",
     "anomaly_narrative"
    ],
    "title": "LlmIntent",
    "type": "string"
   },
...
   "LlmStatus": {
    "properties": {
     "calls_today": {
      "title": "Calls Today",
      "type": "integer"
     },
     "configured": {
      "enum": [
       "auto",
       "anthropic",
       "openai_compatible",
       "replay",
       "template"
      ],
      "title": "Configured",
      "type": "string"
     },
     "daily_cap": {
      "title": "Daily Cap",
      "type": "integer"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "key_configured": {
      "title": "Key Configured",
      "type": "boolean"
     },
     "last_error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Error"
     },
     "last_error_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Error At"
     },
     "live": {
      "title": "Live",
      "type": "boolean"
     },
     "mode": {
      "enum": [
       "anthropic",
       "openai_compatible",
       "replay",
       "template"
      ],
      "title": "Mode",
      "type": "string"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "replay_entries": {
      "title": "Replay Entries",
      "type": "integer"
     },
     "user_calls_per_min": {
      "title": "User Calls Per Min",
      "type": "integer"
     }
    },
    "required": [
     "configured",
     "mode",
     "live",
     "model",
     "key_configured",
     "replay_entries",
     "calls_today",
     "daily_cap",
     "user_calls_per_min",
     "last_error",
     "last_error_at",
     "generated_at"
    ],
    "title": "LlmStatus",
    "type": "object"
   },
   "LlmText": {
    "description": "Generated wording. Numbers inside `text` are verified against the evidence pack.",
    "properties": {
     "advisory": {
      "const": true,
      "default": true,
      "title": "Advisory",
      "type": "boolean"
     },
     "cached": {
      "title": "Cached",
      "type": "boolean"
     },
     "cited_factors": {
      "items": {
       "type": "string"
      },
      "title": "Cited Factors",
      "type": "array"
     },
     "evidence_hash": {
      "title": "Evidence Hash",
      "type": "string"
     },
     "fallback_reason": {
      "anyOf": [
       {
        "enum": [
         "rate_limited",
         "daily_cap",
         "timeout",
         "error",
         "numbers_fail",
         "schema_fail",
         "injection",
         "replay_miss",
         "not_configured"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fallback Reason"
     },
     "generated_at": {
      "format": "date-time",
      "title": "Generated At",
      "type": "string"
     },
     "generated_by": {
      "$ref": "#/components/schemas/GeneratedBy"
     },
     "guard_result": {
      "$ref": "#/components/schemas/GuardResult"
     },
     "intent": {
      "$ref": "#/components/schemas/LlmIntent"
     },
     "lang": {
      "$ref": "#/components/schemas/Lang"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "model_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model Version"
     },
     "provider": {
      "title": "Provider",
      "type": "string"
     },
     "template_text": {
      "title": "Template Text",
      "type": "string"
     },
     "text": {
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "intent",
     "text",
     "lang",
     "cited_factors",
     "generated_by",
     "provider",
     "model",
     "guard_result",
     "fallback_reason",
     "template_text",
     "cached",
     "evidence_hash",
     "model_version",
     "generated_at"
    ],
    "title": "LlmText",
    "type": "object"
   },
...
   "NarrateIn": {
    "properties": {
     "agent_id": {
      "maximum": 2147483647.0,
      "minimum": 1.0,
      "title": "Agent Id",
      "type": "integer"
     },
     "lang": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Lang"
       },
       {
        "type": "null"
       }
      ]
     },
     "target": {
      "$ref": "#/components/schemas/FloatType",
      "default": "cash"
     }
    },
    "required": [
     "agent_id"
    ],
    "title": "NarrateIn",
    "type": "object"
   },
   "NotificationItem": {
    "description": "Text is built client-side from `title_key` + `params` (bn/en i18n).",
    "properties": {
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "entity_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Entity Id"
     },
     "entity_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Entity Type"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "params": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        },
        {
         "type": "number"
        },
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ]
      },
      "title": "Params",
      "type": "object"
     },
     "read_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Read At"
     },
     "severity": {
      "$ref": "#/components/schemas/NotificationSeverity"
     },
     "title_key": {
      "title": "Title Key",
      "type": "string"
     },
     "type": {
      "$ref": "#/components/schemas/NotificationType"
     }
    },
    "required": [
     "id",
     "type",
     "severity",
     "title_key",
     "params",
     "entity_type",
     "entity_id",
     "read_at",
     "created_at"
    ],
    "title": "NotificationItem",
    "type": "object"
   },
   "NotificationPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/NotificationItem"
      },
      "title": "Items",
      "type": "array"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     },
     "unread_count": {
      "title": "Unread Count",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "unread_count",
     "page",
     "page_size"
    ],
    "title": "NotificationPage",
    "type": "object"
   },
   "NotificationSeverity": {
    "enum": [
     "info",
     "warning",
     "critical"
    ],
    "title": "NotificationSeverity",
    "type": "string"
   },
   "NotificationType": {
    "enum": [
     "risk_change",
     "swap_offer",
     "swap_decision",
     "anomaly",
     "system"
    ],
    "title": "NotificationType",
    "type": "string"
   },
...
   "ReadAllResult": {
    "properties": {
     "updated": {
      "title": "Updated",
      "type": "integer"
     }
    },
    "required": [
     "updated"
    ],
    "title": "ReadAllResult",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "AgentPulse AI",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/agents/{agent_id}/briefing": {
   "get": {
    "description": "Morning briefing for one agent: risk, main reason, suggested action (advisory).",
    "operationId": "agent_briefing_api_v1_agents__agent_id__briefing_get",
    "parameters": [
     {
      "in": "path",
      "name": "agent_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Agent Id",
       "type": "integer"
      }
     },
     {
      "description": "bn or en; default: user's language",
      "in": "query",
      "name": "lang",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/Lang"
        },
        {
         "type": "null"
        }
       ],
       "description": "bn or en; default: user's language",
       "title": "Lang"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmText"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Agent Briefing",
    "tags": [
     "llm"
    ]
   }
  },
...
  "/api/v1/anomalies/{anomaly_id}/narrative": {
   "get": {
    "description": "Neutral investigation note for one flag, from its peer evidence only.",
    "operationId": "anomaly_narrative_api_v1_anomalies__anomaly_id__narrative_get",
    "parameters": [
     {
      "in": "path",
      "name": "anomaly_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Anomaly Id",
       "type": "integer"
      }
     },
     {
      "description": "bn or en; default: user's language",
      "in": "query",
      "name": "lang",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/Lang"
        },
        {
         "type": "null"
        }
       ],
       "description": "bn or en; default: user's language",
       "title": "Lang"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmText"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Anomaly Narrative",
    "tags": [
     "llm"
    ]
   }
  },
...
  "/api/v1/distributor/briefing": {
   "get": {
    "description": "Daily briefing over the caller's agents (distributor: own; admin: all).",
    "operationId": "distributor_briefing_api_v1_distributor_briefing_get",
    "parameters": [
     {
      "description": "bn or en; default: user's language",
      "in": "query",
      "name": "lang",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/Lang"
        },
        {
         "type": "null"
        }
       ],
       "description": "bn or en; default: user's language",
       "title": "Lang"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmText"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Distributor Briefing",
    "tags": [
     "llm"
    ]
   }
  },
...
  "/api/v1/explanations/narrate": {
   "post": {
    "description": "SHAP template reasons for one float, reworded by the LLM (template on any failure).",
    "operationId": "narrate_api_v1_explanations_narrate_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/NarrateIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmText"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Narrate",
    "tags": [
     "llm"
    ]
   }
  },
...
  "/api/v1/llm/status": {
   "get": {
    "description": "Configured and effective provider, today's live calls vs the cap, last error code.",
    "operationId": "get_llm_status_api_v1_llm_status_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LlmStatus"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Llm Status",
    "tags": [
     "llm"
    ]
   }
  },
...
  "/api/v1/notifications": {
   "get": {
    "description": "The caller's own notifications, newest first, with the unread count for the bell.",
    "operationId": "list_notifications_api_v1_notifications_get",
    "parameters": [
     {
      "description": "true: unread only; false: read only",
      "in": "query",
      "name": "unread",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "description": "true: unread only; false: read only",
       "title": "Unread"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "maximum": 10000,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 100,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Notifications",
    "tags": [
     "notifications"
    ]
   }
  },
  "/api/v1/notifications/read-all": {
   "post": {
    "operationId": "read_all_api_v1_notifications_read_all_post",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReadAllResult"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Read All",
    "tags": [
     "notifications"
    ]
   }
  },
  "/api/v1/notifications/{notification_id}/read": {
   "post": {
    "operationId": "read_one_api_v1_notifications__notification_id__read_post",
    "parameters": [
     {
      "in": "path",
      "name": "notification_id",
      "required": true,
      "schema": {
       "maximum": 2147483647,
       "minimum": 1,
       "title": "Notification Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NotificationItem"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Read One",
    "tags": [
     "notifications"
    ]
   }
  },
...
 }
}